third_party/WebKit/LayoutTests/external/wpt/cssom-view/HTMLBody-ScrollArea_quirksmode-expected.txt

This is a testharness.js-based test.
PASS Ensure that body element is loaded. 
PASS Ensure that style.overflowY can be set properly. 
PASS document.compatMode should be BackCompat in quirks. 
FAIL document.scrollingElement should be body element in quirks. assert_equals: scrollingElement in quirks mode should default to body element. expected Element node <body id="thebody" style="overflow-y: scroll;">
    <div ... but got null
PASS scrollingElement in quirks should be null when body is potentially scrollable. 
FAIL scrollingElement in quirks should be body if any of document and body has a visible overflow. assert_equals: expected Element node <body id="thebody" style="overflow-y: scroll;">
    <div ... but got null
PASS When body potentially scrollable, document.body.scrollHeight changes when changing the height of the body content in quirks. 
FAIL When body not potentially scrollable, document.body.scrollHeight always equals window.innerHeight in quirks. (cond. visible, scroll) assert_equals: expected 600 but got 240
PASS When body not potentially scrollable, document.body.scrollHeight always equals window.innerHeight in quirks. (cond. scroll, visible) 
PASS When body not potentially scrollable, document.body.scrollHeight always equals window.innerHeight in quirks. (cond. visible, visible) 
Harness: the test ran to completion.
